vesta-dev-master/requirements.txt
line-solver==2.0.33.3
enum_tools
jpype1
numpy
pandas
matplotlib
networkx
#openbox
#nevergrad
#shap # used by openbox
#lightgbm # used by openbox
#opytimizer
#scikit-opt
#websocket
smt==2.7.0
jenn==1.0.8
pyDOE3
scipy
bayesian-optimization===1.4.3 
nevergrad==1.0.3